requirements.txt
anaconda==0.0.1.1
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.9
future==0.18.2
gunicorn==20.0.4
idna==2.10
jieba==0.42.1
line-bot-sdk==1.16.0
lxml==4.5.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
twder==0.1.4
urllib3==1.25.10
whitenoise==5.1.0
